Sprint 1: heap-corruption, flak-rand, texture-truncation, CRLF fixes
Correctness hardening sweep (Wave A), all portable / no behavior change on Windows:

- new[]/delete -> delete[] in 8 network message dtors (falclib/msgsrc):
  falconflightplanmsg, sendimage, sendevalmsg, simdirtydatamsg, sendvcmsg,
  requestcampaigndata, sendcampaignmsg, sendpersistantlist. Array buffers
  (new uchar[]/VU_BYTE[]) were freed with scalar delete = heap corruption.
- shells.cpp flak: rand()/32767.0f -> rand()/(float)RAND_MAX (2 active sites).
  glibc rand() is 0..2^31 so the old divisor made air-burst hit probability
  ~65000x too low (enemy AAA ineffective). RAND_MAX==32767 on MSVC so Windows
  is unchanged.
- gmcomposit.cpp: SelectTexture1((UInt)texHandle) -> (intptr_t) (3 sites);
  TextureHandle* was truncated to 32 bits (matches existing line-699 fix).
- CRLF: strip trailing \r/\n after fgets in entity.cpp rack parser and the two
  realweather METAR readers (last token / per-field strcpy were polluted).

+++ src/graphics/renderer/gmcomposit.cpp
@@ -413,7 +413,7 @@ void RenderGMComposite::DrawComposite(Tpoint *center, float platformHdg)
         }
 
         context.RestoreState(STATE_TEXTURE);
-        context.SelectTexture1((UInt) rTexHandle);
+        context.SelectTexture1((intptr_t) rTexHandle);
         ClipAndDraw2DFan(vertArray, num);
     }
 
@@ -479,7 +479,7 @@ void RenderGMComposite::DrawComposite(Tpoint *center, float platformHdg)
         }
 
         context.RestoreState(STATE_TEXTURE);
-        context.SelectTexture1((UInt) lTexHandle);
+        context.SelectTexture1((intptr_t) lTexHandle);
         ClipAndDraw2DFan(vertArray, num);
     }
 
@@ -848,7 +848,7 @@ void RenderGMComposite::DebugDrawLeftTexture(Render2D *renderer)
 
 
     renderer->context.RestoreState(STATE_TEXTURE);
-    renderer->context.SelectTexture1((UInt) lTexHandle);
+    renderer->context.SelectTexture1((intptr_t) lTexHandle);
 
     v0.x = 1.0f;
     v0.y = 1.0f;
